+++ apidoc/changelog.rst
@@ -5,6 +5,15 @@
 Release Notes and Changelog
 ===========================
 
+
+Bugfix Release 0.8.2
+=====================
+
+* Added backward compatibility wrappers and deprecation warnings to some of the API changes.
+* Fixed "FileCheckerThread seems to fail on eggs" (Issue #87)
+* Fixed "Bottle.get_url() does not return correct path when SCRIPT_NAME is set." (Issue #83)
+
+
 Release 0.8
 ===========